+++ src/types/posts/post.ts
@@ -1,0 +1,20 @@
+import { ContentBlocks } from '@typings/contentBlocks';
+
+export interface Post {
+  id: number;
+  title: string;
+  slug: string;
+  date: string;
+  authorId: number;
+  categories?: number[];
+  tags?: number[];
+  excerpt?: string;
+  content?: ContentBlocks;
+  image?: string;
+  relatedPosts?: number[];
+  postHeader: {
+    layout: 'outside-above' | 'outside-below' | 'split-narrow' | 'split-wide' | 'split-full' | 'overlay-narrow' | 'overlay-wide' | 'overlay-full';
+    besideSidebar: boolean;
+  };
+  postSidebar: 'left' | 'right' | 'hidden';
+}