+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Header components to TypeScript
 - Footer components to TypeScript
+- Form components (checkbox, password, select, textarea) to TypeScript
 
 ## [0.0.0] - 2025-11-21